refactor: move dataclasses to skeliner/dataclass.py
